Split Snapshot tests into separate files
